% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvsample.R
\name{rvsample}
\alias{rvsample}
\title{4800shilin}
\usage{
rvsample(n, pdf, lb, ub, c)
}
\arguments{
\item{n}{the number of samples}

\item{pdf}{a function that is the pdf of the random variable that you wish to sample from}

\item{lb}{a numeric that is the lower bound of the random variable you wish to sample from}

\item{ub}{a numeric that is the upper bound of the random variable you wish to sample from m, so P(a <= X <= b) = 1.}

\item{c}{a numeric that is such that f(x) <= C for all values of x}
}
\value{
a random sample of size n from the rv with pdf provided to the function, using rejection sampling
}
\description{
what does the function do: implements sampling from a continuous random variable.
}
\details{
what does the function do: implements sampling from a continuous random variable.
}
